api/requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==4.2.10
django-channels==0.7.0
django-minio-storage==0.5.7
django-redis==5.4.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
kombu==5.3.5
minio==7.2.4
msgpack==1.0.7
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pyOpenSSL==24.0.0
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-django==4.8.0
pytz==2024.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.10.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
